Split outside_connections/tests_detection.rs into smaller modules
Split crates/simulation/src/outside_connections/tests_detection.rs (518 lines)
into two focused test modules:
- tests_detection_edge_highway.rs (228 lines): edge detection + highway tests
- tests_detection_services.rs (297 lines): railway, seaport, airport, capacity tests
